requirements.txt
Flask
gunicorn
torch
torchvision
beautifulsoup4
requests
Pillow